Player Manager displasy stuff works.
Also some namespace changing for displays, my test scene.
