+++ Program.cs
@@ -1,12 +1,12 @@
 ﻿using TowerOffense;
 using TowerOffense.Scenes.Example;
 using TowerOffense.Scenes.Title;
-using TowerOffense.Scenes.HpTest;
+using TowerOffense.Scenes.Test;
 
 public class Program {
     private static void Main(string[] args) {
         using var game = new TOGame();
-        TOGame.Scenes.PushScene<TitleScene>();
+        TOGame.Scenes.PushScene<DisplayTestScene>();
         game.Run();
     }
 }
