+++ TowerOffense.Objects/Displays/HPDisplay.cs
@@ -18,7 +18,7 @@ namespace TowerOffense.Displays {
         }
         public override void Render(GameTime gameTime){
             Vector2 textMiddlePoint = _font.MeasureString(_text) / 2;
-            Vector2 position = new Vector2(48,48); // some work to center this would be cool
+            Vector2 position = new Vector2(0,this.InnerSize.Y/2); // some work to center this would be cool
             DrawString(_font, _text, position, Color.Red);
             base.Render(gameTime);
         }
